composables: map: extract the custom tile provider layer factory
Move the URL and archive layer construction out of the layer-control
registry into its own composable, so a map that selects a single provider
without a layer control can build the same layers instead of duplicating
the archive download-and-cache path.
